Hospital Data: Use vuex to control the data flow
