+++ src/stores/index.js
@@ -2,7 +2,7 @@ import Vuex from 'vuex';
 import Vue from 'vue';
 import country from './modules/country';
 import global from './modules/global';
-
+import hospitals from './modules/hospitals';
 // Load Vuex
 Vue.use(Vuex);
 
@@ -10,6 +10,7 @@ Vue.use(Vuex);
 export default new Vuex.Store({
   modules: {
     country,
-    global
+    global,
+    hospitals
   }
 });
